requirements.txt
bidict==0.22.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-SocketIO==5.2.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
PyPDF2==2.11.1
python-docx==0.8.11
python-engineio==4.3.4
python-socketio==5.7.2
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.10.0
